+++ Realization/frontend/czechidm-app/src/Index.js
@@ -190,7 +190,17 @@ function reduxQueue({ dispatch, getState }) {
   }
   //
   return next => action => {
-    const { queue: key, callback, id } = action || {};
+    const { queue: key, callback, id, type } = action || {};
+    if (type === 'RECEIVE_LOGIN') {
+      // its needed to clear the queues after login (queues can contain incomplete requests)
+      for (const queueKey in running) {
+        if (!running.hasOwnProperty(queueKey)) {
+          continue;
+        }
+        running[queueKey] = new Immutable.Set();
+      }
+    }
+    //
     if (key) {
       if (typeof callback !== 'function') {
         throw new Error('Queued actions must have a <callback> property');